requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0